Moved the input layer up in the file hierarchy to the io directory.
