+++ include/lbann/callbacks/debug_io.hpp
@@ -32,7 +32,7 @@
 #include <chrono>
 #include <vector>
 #include "lbann/callbacks/callback.hpp"
-#include "lbann/layers/io/input/input_layer.hpp"
+#include "lbann/layers/io/input_layer.hpp"
 
 namespace lbann {
 namespace callback {
