+++ include/lbann/layers/io/CMakeLists.txt
@@ -1,9 +1,7 @@
 # Add the headers for this directory
 set_full_path(THIS_DIR_HEADERS
+  input_layer.hpp
   )
-
-# Add the subdirectories
-add_subdirectory(input)
 
 # Propagate the files up the tree
 set(HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
