--- include/lbann/layers/io/input/CMakeLists.txt
@@ -1,7 +1,0 @@
-# Add the headers for this directory
-set_full_path(THIS_DIR_HEADERS
-  input_layer.hpp
-  )
-
-# Propagate the files up the tree
-set(HEADERS "${HEADERS}" "${THIS_DIR_HEADERS}" PARENT_SCOPE)
