+++ include/lbann/lbann.hpp
@@ -107,7 +107,7 @@
 #include "lbann/layers/regularizers/instance_norm.hpp"
 
 /// Input layer
-#include "lbann/layers/io/input/input_layer.hpp"
+#include "lbann/layers/io/input_layer.hpp"
 
 /// Miscellaneous layers
 #include "lbann/layers/misc/covariance.hpp"
