+++ src/layers/CMakeLists.txt
@@ -14,7 +14,7 @@ endif ()
 # Add the subdirectories
 add_subdirectory(activations)
 add_subdirectory(image)
-add_subdirectory(io/input)
+add_subdirectory(io)
 add_subdirectory(learning)
 add_subdirectory(loss)
 add_subdirectory(math)
