+++ src/proto/factories/layer_factory.cpp
@@ -38,8 +38,7 @@
 #include "lbann/layers/activations/log_softmax.hpp"
 #include "lbann/layers/activations/softmax.hpp"
 #include "lbann/layers/image/bilinear_resize.hpp"
-#include "lbann/layers/io/input/input_layer.hpp"
-#include "lbann/layers/io/io_layer.hpp"
+#include "lbann/layers/io/input_layer.hpp"
 #include "lbann/layers/learning/base_convolution.hpp"
 #include "lbann/layers/learning/channelwise_fully_connected.hpp"
 #include "lbann/layers/learning/channelwise_scale_bias.hpp"
